Refactor field export for more generic use
