+++ src/main/java/dk/kb/netarchivesuite/solrwayback/export/ContentStreams.java
@@ -15,21 +15,24 @@
 package dk.kb.netarchivesuite.solrwayback.export;
 
 import com.google.common.base.Functions;
-import dk.kb.netarchivesuite.solrwayback.facade.Facade;
 import dk.kb.netarchivesuite.solrwayback.service.dto.ArcEntryDescriptor;
 import dk.kb.netarchivesuite.solrwayback.solr.SolrGenericStreaming;
 import dk.kb.netarchivesuite.solrwayback.util.CollectionUtils;
+import dk.kb.netarchivesuite.solrwayback.util.JsonUtils;
 import dk.kb.netarchivesuite.solrwayback.util.Processing;
 import dk.kb.netarchivesuite.solrwayback.util.SolrUtils;
+import dk.kb.netarchivesuite.solrwayback.util.StreamBridge;
 import org.apache.solr.common.SolrDocument;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.util.ArrayList;
+import java.io.IOException;
+import java.io.InputStream;
 import java.util.Arrays;
 import java.util.List;
+import java.util.Locale;
 import java.util.concurrent.Callable;
-import java.util.stream.Collectors;
+import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.stream.Stream;
 
 /**
@@ -37,6 +40,14 @@ import java.util.stream.Stream;
  */
 public class ContentStreams {
     private static final Logger log = LoggerFactory.getLogger(ContentStreams.class);
+
+    public enum FORMAT {
+        /** Comma Separated Values. */
+        csv,
+        /** Standard JSON, with the documents in a JSON array. */
+        json,
+        /** <a href="https://jsonlines.org/">JSON-Lines</a> with a single JSON block per line. */
+        jsonl}
 
     /**
      * Searches images and webpages matching the given searchText. For webpages, linked images are resolved and
@@ -69,7 +80,7 @@ public class ContentStreams {
     }
 
     /**
-     * Creates a callable delivering up to maxImages {@link ArcEntryDescriptor}s for images listed in
+     * Create a callable delivering up to maxImages {@link ArcEntryDescriptor}s for images listed in
      * {@code links_images} for the given htmlPage. The images are searched using
      * {@link SolrGenericStreaming.SRequest#timeProximityDeduplication(String, String)} meaning that only one
      * instance of a given image URL is returned, with preference to the one nearest in time to the htmlPage.
@@ -101,4 +112,79 @@ public class ContentStreams {
                 map(SolrUtils::solrDocument2ArcEntryDescriptor);
     }
 
+    /**
+     * Use a Thread to write the {@code fields} from the given {@code docs} in the given {@code format} to the
+     * returned InputStream.
+     *
+     * Supported formats are defined in {@link FORMAT}.
+     * @param docs   Solr documents with the stated {@code fields}.
+     * @param fields the fields to write. Some {@code format}s ignore these and write all fields in the {@code docs}.
+     * @param format the format to write the data. See {@link FORMAT}.
+     * @throws IOException if the content could not be written.
+     */
+    public static InputStream deliver(Stream<SolrDocument> docs, String fields, String format) throws IOException {
+        FORMAT realFormat = FORMAT.valueOf(format.toLowerCase(Locale.ROOT));
+
+        return StreamBridge.outputToInputSafe(out -> {
+            switch (realFormat) {
+                case csv: {
+                    writeCSV(docs, fields, out);
+                    break;
+                }
+                case json: {
+                    writeJSON(docs, out);
+                    break;
+                }
+                case jsonl: {
+                    writeJSONLines(docs, out);
+                    break;
+                }
+                default:
+                    throw new UnsupportedOperationException("The format '" + format + "' is not supported");
+            }
+        });
+    }
+
+    /**
+     * Write the given SolrDocuments as <a href="https://jsonlines.org/">JSON-Lines</a>: One JSON block for each
+     * document, with one block per line.
+     * @param docs a Stream of Solr documents.
+     * @param out where to write the output.
+     */
+    private static void writeJSONLines(Stream<SolrDocument> docs, StreamBridge.SafeOutputStream out) {
+        docs.map(JsonUtils::toJSON).
+                forEach(out::writeln);
+    }
+
+    /**
+     * Write the given SolrDocuments as a single JSON array containing one JSON block for each document.
+     * @param docs a Stream of Solr documents.
+     * @param out where to write the output.
+     */
+    private static void writeJSON(Stream<SolrDocument> docs, StreamBridge.SafeOutputStream out) {
+        AtomicBoolean hasWritten = new AtomicBoolean(false);
+        out.writeln("[");
+        docs.map(JsonUtils::toJSON).
+                peek(json -> {
+                    if (!hasWritten.get()) {
+                        hasWritten.set(true);
+                    } else {
+                        out.writeln(",");
+                    }
+                }).
+                forEach(out::write);
+        out.writeln("\n]");
+    }
+
+    /**
+     * Write the given SolrDocuments as Comma Separated Values.
+     * First line acts as header and lists the field names.
+     * @param docs a Stream of Solr documents.
+     * @param out where to write the output.
+     */
+    private static void writeCSV(Stream<SolrDocument> docs, String fields, StreamBridge.SafeOutputStream out) {
+        GenerateCSV csvMapper = new GenerateCSV(fields.split(", *"));
+        docs.map(csvMapper::toCVSLine).
+                forEach(out::write);
+    }
 }
